Fix yet another Svace static analyzer warnings (#1372)

+++ src/libdrakvuf/libdrakvuf.h
@@ -408,8 +408,8 @@ addr_t drakvuf_get_obj_by_handle(drakvuf_t drakvuf,
 
 os_t drakvuf_get_os_type(drakvuf_t drakvuf) NOEXCEPT;
 page_mode_t drakvuf_get_page_mode(drakvuf_t drakvuf) NOEXCEPT;
-int drakvuf_get_address_width(drakvuf_t drakvuf) NOEXCEPT;
-int drakvuf_get_process_address_width(drakvuf_t drakvuf,
+size_t drakvuf_get_address_width(drakvuf_t drakvuf) NOEXCEPT;
+size_t drakvuf_get_process_address_width(drakvuf_t drakvuf,
     drakvuf_trap_info_t* info) NOEXCEPT;
 int drakvuf_read_addr(drakvuf_t drakvuf, drakvuf_trap_info_t* info,
     const access_context_t* ctx, addr_t* addr) NOEXCEPT;
