+++ src/plugins/filetracer/win.cpp
@@ -739,7 +739,7 @@ win_filetracer::win_filetracer(drakvuf_t drakvuf, output_format_t output)
     : format{output}
     , offsets(new size_t[__OFFSET_MAX])
 {
-    int addr_size = drakvuf_get_address_width(drakvuf); // 4 or 8 (bytes)
+    uint64_t addr_size = drakvuf_get_address_width(drakvuf); // 4 or 8 (bytes)
 
     if ( !drakvuf_get_kernel_struct_members_array_rva(drakvuf, offset_names, __OFFSET_MAX, offsets) )
         throw -1;
